feat: Extract HabitsTable into a component and start mock db implementation
