+++ src/app/habits/types.ts
@@ -1,0 +1,6 @@
+export type Mark = "" | "•" | "X";
+
+export type Habit = {
+  name: string;
+  marks: Mark[];
+};
